Moving lerna tasks out into seperate file and fixing tests
