+++ gulp-tasks/build.js
@@ -19,8 +19,7 @@ const chalk = require('chalk');
 const fse = require('fs-extra');
 const gulp = require('gulp');
 const path = require('path');
-const runSequence = require('run-sequence');
-const {taskHarness, buildJSBundle, lernaWrapper} = require('../utils/build');
+const {taskHarness, buildJSBundle} = require('../utils/build');
 const commonjs = require('rollup-plugin-commonjs');
 const resolve = require('rollup-plugin-node-resolve');
 
@@ -48,28 +47,6 @@ const buildPackage = (projectPath) => {
     .then(() => fse.copy(path.join(__dirname, '..', 'LICENSE'),
       path.join(projectPath, 'LICENSE')))
     .then(() => printBuildTime(`${(Date.now() - startTime) / 1000}s`));
-};
-
-/**
- * Updates the fields in package.json that contain version string to match
- * the latest version from lerna.json.
- *
- * @param {String} projectPath The path to a project directory.
- * @return {Promise} Resolves if updating succeeds, and rejects if it fails.
- */
-const updateVersionedBundles = (projectPath) => {
-  const packageJsonPath = path.join(projectPath, 'package.json');
-
-  return fse.readJson(packageJsonPath).then((pkg) => {
-    const regexp = /v\d+\.\d+\.\d+/;
-    for (let field of ['main', 'module']) {
-      if (field in pkg) {
-        pkg[field] = pkg[field].replace(regexp, `v${pkg.version}`);
-      }
-    }
-
-    return fse.writeJson(packageJsonPath, pkg, {spaces: 2});
-  });
 };
 
 gulp.task('build:shared', () => {
@@ -117,53 +94,4 @@ gulp.task('build', () => {
 gulp.task('build:watch', ['build'], (unusedCallback) => {
   gulp.watch(`packages/${global.projectOrStar}/src/**/*`, ['build']);
   gulp.watch(`lib/**/*`, ['build']);
-});
-
-gulp.task('lerna-bootstrap', () => {
-  return lernaWrapper('bootstrap');
-});
-
-gulp.task('lerna-bootstrap-scoped', () => {
-  return lernaWrapper('bootstrap', '--include-filtered-dependencies', '--scope',
-    global.projectOrStar);
-});
-
-/**
- * Helper task, used only within lerna-publish.
- */
-gulp.task('_lerna-publish-dry-run', () => {
-  return lernaWrapper('publish', '--skip-npm', '--skip-git');
-});
-
-/**
- * Helper task, used only within lerna-publish.
- */
-gulp.task('_update-versioned-bundles', () => {
-  return taskHarness(updateVersionedBundles, global.projectOrStar);
-});
-
-/**
- * Helper task, used only within lerna-publish.
- */
-gulp.task('_lerna-publish-repo-version', () => {
-  return fse.readJson('lerna.json').then((lernaConfig) => {
-    return lernaWrapper('publish', '--yes', '--repo-version',
-      lernaConfig.version);
-  });
-});
-
-/**
- * This is the task you should use to publish a new release of updated
- * modules to npm.
- */
-gulp.task('lerna-publish', (callback) => {
-  runSequence(
-    'lerna-bootstrap',
-    'test:dev',
-    'test:prod',
-    '_lerna-publish-dry-run',
-    '_update-versioned-bundles',
-    '_lerna-publish-repo-version',
-    callback
-  );
 });
