+++ gulpfile.js
@@ -35,6 +35,7 @@ global.projectOrStar = options.project || '*';
 global.cliOptions = options;
 
 require('./gulp-tasks/lint.js');
+require('./gulp-tasks/lerna.js');
 require('./gulp-tasks/build.js');
 require('./gulp-tasks/test.js');
 require('./gulp-tasks/serve.js');
