+++ lerna.json
@@ -1,5 +1,6 @@
 {
   "lerna": "2.0.0-rc.5",
+  "version": "1.0.1",
   "packages": [
     "packages/*"
   ]
